bin/haproxy_config/requirements.txt
docker-py
watchdog